Switch to 'proper' voxels
Switch to importing PRIMITIVE_POINTS instead of PRIMITIVE_TRIANGLES
Add a shader for voxels
Add supporting script to adjust the point_size and track viewport size
changes.

This version works, but it's still rough, e.g. if you tune the
point_size to the center of the screen, the voxels separate when you see
the model from the corner of your eye, etc. Point_size needs to somehow
be adjusted based on distance from the center

Usage:

Import your .vox files with the modified plugin.
Set the Point Size of your imported voxel mesh.
NOTE: my voxel meshes are scaled to 0.026, which requires a 'point size'
of 24, not sure what would look good at other sizes.