Stop using arrays for voxel data because they are actually slower
(Could be faster if the voxel data was not sparse, but it almost always
is)

Fix voxel mesh: issues when re-importing, shader material point size was
not re-set after re-importing the vox mesh.

Make all the sub classes references so they will be deleted after
import.

Add a class that handles lookups of the voxel data.

Add debug timing printout that can be enabled.